requirements.txt
requests>=2.32.0
PyYAML>=6.0.1
Pillow>=10.4.0
playwright>=1.53.0
